ci: ensure type ignores in langchain tests run across all versions of langchain (#460)
The issue is that our type ignores are required on the old version of LangChain, but not required on the new version of LangChain. If we include the type ignore, the new version of LangChain complains about an usused type ignore. Found the solution [here](https://github.com/python/mypy/issues/8823#issuecomment-1556288015).