fix(js/langchain): Ensure esm builds are emitted correctly
We had an incorrect module and moduleResolution setting within our tsconfig.esm.json and tsconfig.esnext.json files.

Because the package is commonjs by default, and is missing {type: module} in the root package.json, tsc emits commonjs no matter what. Changing to preserve + bundler ensures that the code we write (esm) is emitted as-is for the esm tsconfig target. A follow up pass with tsc-alias rewrites the imports to ensure they have file extensions.

Esnext has been adapted similarly, but without the tsc-alias pass.